Made auth support generic
