+++ example.js
@@ -11,6 +11,42 @@ function build (opts) {
     .register(require('./fastify-auth'))
     .after(routes)
 
+  fastify.decorate('verifyJWTandLevel', verifyJWTandLevel)
+
+  function verifyJWTandLevel (request, reply, done) {
+    const jwt = this.jwt
+    const level = this.level
+
+    if (!request.req.headers['auth']) {
+      return done(null, false, 'Missing token header')
+    }
+
+    jwt.verify(request.req.headers['auth'], onVerify)
+
+    function onVerify (err, decoded) {
+      if (err || !decoded.user || !decoded.password) {
+        return done(null, false, 'Token not valid')
+      }
+
+      level.get(decoded.user, onUser)
+
+      function onUser (err, password) {
+        if (err) {
+          if (err.notFound) {
+            return done(null, false, 'Token not valid')
+          }
+          return done(err)
+        }
+
+        if (!password || password !== decoded.password) {
+          return done(null, false, 'Token not valid')
+        }
+
+        done(null, true)
+      }
+    }
+  }
+
   function routes () {
     fastify.route({
       method: 'POST',
@@ -54,6 +90,7 @@ function build (opts) {
     fastify.route({
       method: 'GET',
       url: '/auth',
+      auth: [fastify.verifyJWTandLevel.bind(fastify)],
       beforeHandler: fastify.auth.bind(fastify),
       handler: (req, reply) => {
         req.log.info('Auth route')
